Move the retrieval logic from endpoints to the controllers
