+++ src/app/modules/collection/types.ts
@@ -1,3 +1,6 @@
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { ModuleEndpointContext } from 'lisk-framework';
+
 export interface Collection {
   name: string;
   releaseYear: string;
@@ -52,4 +55,9 @@ export interface SetAttributesCommandParams {
   coArtists: string[];
   type: number;
   collectionID: Buffer;
+}
+
+export interface Store<Entity> {
+  get:  (context: ModuleEndpointContext, key: Buffer) => Promise<Entity>;
+  has: (context: ModuleEndpointContext, key: Buffer) => Promise<boolean>;
 }
